rsx: Minor fixes - vk: Clear dirty textures before copying 'old contents' in case the old data does not fill the new region - rsx: Properly decode border color - seems to be in BGRA format - vk: better approximation of border color to better choose between the presets - vk: Individually clear color images outside render pass and without scissor - vk: Fix renderpass selection for clear overlay pass - vk: Include scissor region when emulating clear mask
NOTES:
- vk: Completely avoid using vkClearXXXXimage - its 'broken' on nvidia drivers
  Spec is vague about the function so its not an actual bug
  ClearAttachment is clearly defined as bypassing bound state which works correctly
- TODO: Implement memory sampling to simulate loading precleared memory if cell used memset to preinitialize the framebuffer
  Autoclear depth to 1|255 and color to 0 is hacky!

+++ rpcs3/Emu/RSX/VK/VKFormats.h
@@ -11,7 +11,7 @@ namespace vk
 	};
 
 	gpu_formats_support get_optimal_tiling_supported_formats(VkPhysicalDevice physical_device);
-	VkBorderColor get_border_color(u8 color);
+	VkBorderColor get_border_color(u32 color);
 
 	VkFormat get_compatible_depth_surface_format(const gpu_formats_support &support, rsx::surface_depth_format format);
 	VkFormat get_compatible_sampler_format(u32 format);
